Streamlit Application/requirements.txt
pillow
numpy
tensorflow
keras
scikit-learn
matplotlib
pandas
scikit-image
seaborn
scipy